+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/circuit/CircuitFirExtension.kt
@@ -64,6 +64,7 @@ import org.jetbrains.kotlin.fir.symbols.impl.FirNamedFunctionSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirRegularClassSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirValueParameterSymbol
 import org.jetbrains.kotlin.fir.types.ConeKotlinType
+import org.jetbrains.kotlin.fir.types.FirImplicitTypeRef
 import org.jetbrains.kotlin.fir.types.FirResolvedTypeRef
 import org.jetbrains.kotlin.fir.types.builder.buildResolvedTypeRef
 import org.jetbrains.kotlin.fir.types.classId
@@ -475,7 +476,14 @@ public class CircuitFirExtension(session: FirSession, compatContext: CompatConte
     return computedTargets.getOrPut(factoryClassId) {
       val function = functionFactoryClassIds[factoryClassId] ?: return@getOrPut null
       val typeResolver = typeResolverFactory.create(function) ?: return@getOrPut null
-      val returnType = typeResolver.resolveType(function.fir.returnTypeRef)
+      val returnTypeRef = function.fir.returnTypeRef
+      val returnType =
+        if (returnTypeRef is FirImplicitTypeRef) {
+          // Assume it's Unit/UI. Checker will validate otherwise later
+          session.builtinTypes.unitType.coneType
+        } else {
+          typeResolver.resolveType(returnTypeRef)
+        }
       val factoryType = if (returnType.isUnit) FactoryType.UI else FactoryType.PRESENTER
       computeFactoryTarget(function, factoryClassId, typeResolver, factoryType, returnType)
     }